second pass

+++ src/data/GameContext.cpp
@@ -427,7 +427,7 @@ bool GameContext::SaveLocal() const
 {
     // Commits local achievements to the file
     auto& pLocalStorage = ra::services::ServiceLocator::GetMutable<ra::services::ILocalStorage>();
-    auto pData = pLocalStorage.WriteText(ra::services::StorageItemType::UserAchievements, std::to_wstring(static_cast<unsigned int>(GameId())));
+    auto pData = pLocalStorage.WriteText(ra::services::StorageItemType::UserAchievements, std::to_wstring(GameId()));
     if (pData == nullptr)
         return false;
